@@ -1,9 +1,8 @@
 ﻿using CaseOnline.Azure.WebJobs.Extensions.Mqtt.Config;
 
-namespace CaseOnline.Azure.WebJobs.Extensions.Mqtt.Bindings
+namespace CaseOnline.Azure.WebJobs.Extensions.Mqtt.Bindings;
+
+public interface IMqttConfigurationParser
 {
-    public interface IMqttConfigurationParser
-    {
-        MqttConfiguration Parse(MqttBaseAttribute mqttAttribute);
-    }
+    MqttConfiguration Parse(MqttBaseAttribute mqttAttribute);
 }
