@@ -1,41 +1,39 @@
-﻿using System;
-using System.Runtime.Serialization;
+﻿using System.Runtime.Serialization;
 
-namespace CaseOnline.Azure.WebJobs.Extensions.Mqtt.Bindings
+namespace CaseOnline.Azure.WebJobs.Extensions.Mqtt.Bindings;
+
+/// <summary>
+/// Thrown when an invalid custom config creator is provided.
+/// </summary>
+[Serializable]
+public class InvalidCustomConfigCreatorException : Exception
 {
     /// <summary>
-    /// Thrown when an invalid custom config creator is provided.
+    /// Initializes a new instance of the <see cref="InvalidCustomConfigCreatorException"/> class.
     /// </summary>
-    [Serializable]
-    public class InvalidCustomConfigCreatorException : Exception
+    public InvalidCustomConfigCreatorException()
     {
-        /// <summary>
-        /// Initializes a new instance of the <see cref="InvalidCustomConfigCreatorException"/> class.
-        /// </summary>
-        public InvalidCustomConfigCreatorException()
-        {
-        } 
+    } 
 
-        /// <summary>
-        /// Initializes a new instance of the <see cref="InvalidCustomConfigCreatorException"/> class.
-        /// </summary>
-        /// <param name="message">The error message.</param>
-        public InvalidCustomConfigCreatorException(string message) : base(message)
-        {
-        }
+    /// <summary>
+    /// Initializes a new instance of the <see cref="InvalidCustomConfigCreatorException"/> class.
+    /// </summary>
+    /// <param name="message">The error message.</param>
+    public InvalidCustomConfigCreatorException(string message) : base(message)
+    {
+    }
 
-        /// <summary>
-        /// Initializes a new instance of the <see cref="InvalidCustomConfigCreatorException"/> class.
-        /// </summary>
-        /// <param name="message">The error message.</param>
-        /// <param name="inner">The inner exception.</param>
-        public InvalidCustomConfigCreatorException(string message, Exception inner) : base(message, inner)
-        { 
-        }
+    /// <summary>
+    /// Initializes a new instance of the <see cref="InvalidCustomConfigCreatorException"/> class.
+    /// </summary>
+    /// <param name="message">The error message.</param>
+    /// <param name="inner">The inner exception.</param>
+    public InvalidCustomConfigCreatorException(string message, Exception inner) : base(message, inner)
+    { 
+    }
 
-        protected InvalidCustomConfigCreatorException(SerializationInfo info, StreamingContext context)
-            : base(info, context)
-        {
-        }
+    protected InvalidCustomConfigCreatorException(SerializationInfo info, StreamingContext context)
+        : base(info, context)
+    {
     }
 }
