@@ -1,9 +1,8 @@
 ﻿using MQTTnet.Extensions.ManagedClient;
 
-namespace CaseOnline.Azure.WebJobs.Extensions.Mqtt.Listeners
+namespace CaseOnline.Azure.WebJobs.Extensions.Mqtt.Listeners;
+
+public interface IManagedMqttClientFactory
 {
-    public interface IManagedMqttClientFactory
-    {
-        IManagedMqttClient CreateManagedMqttClient();
-    }
+    IManagedMqttClient CreateManagedMqttClient();
 }
