@@ -1,10 +1,9 @@
-﻿namespace CaseOnline.Azure.WebJobs.Extensions.Mqtt.Messaging
+﻿namespace CaseOnline.Azure.WebJobs.Extensions.Mqtt.Messaging;
+
+public enum MqttRetainHandling
 {
-    public enum MqttRetainHandling
-    {
-        SendAtSubscribe = 0,
-        SendAtSubscribeIfNewSubscriptionOnly = 1,
-        DoNotSendOnSubscribe = 2,
-        NotSet
-    }
+    SendAtSubscribe = 0,
+    SendAtSubscribeIfNewSubscriptionOnly = 1,
+    DoNotSendOnSubscribe = 2,
+    NotSet
 }
