@@ -1,20 +1,17 @@
-﻿using System;
+﻿namespace CaseOnline.Azure.WebJobs.Extensions.Mqtt;
 
-namespace CaseOnline.Azure.WebJobs.Extensions.Mqtt
+public abstract class MqttBaseAttribute : Attribute
 {
-    public abstract class MqttBaseAttribute : Attribute
+    protected MqttBaseAttribute()
     {
-        protected MqttBaseAttribute()
-        {
-        }
-
-        protected MqttBaseAttribute(Type mqttConfigCreatorType)
-        {
-            MqttConfigCreatorType = mqttConfigCreatorType;
-        }
-
-        public string ConnectionString { get; set; }
-
-        public Type MqttConfigCreatorType { get; protected set; }
     }
+
+    protected MqttBaseAttribute(Type mqttConfigCreatorType)
+    {
+        MqttConfigCreatorType = mqttConfigCreatorType;
+    }
+
+    public string ConnectionString { get; set; }
+
+    public Type MqttConfigCreatorType { get; protected set; }
 }
